Move grdcut to src
