+++ pygmt/datasets/earth_relief.py
@@ -6,9 +6,9 @@ The grids are available in various resolutions.
 """
 import xarray as xr
 from pygmt.exceptions import GMTInvalidInput
-from pygmt.gridops import grdcut
 from pygmt.helpers import kwargs_to_strings
 from pygmt.modules import which
+from pygmt.src import grdcut
 
 
 @kwargs_to_strings(region="sequence")
